command_handler/src/requirements.txt
pynacl
pydantic